requirements.txt
tensorflow
pillow
numpy
h5py
streamlit